# marketplace/docs/swagger.yaml
basePath: /api/market/
definitions:
  main.CarbonCredit:
    properties:
      createdAt:
        type: string
      creditsAvailable:
        type: number
      creditsSold:
        type: number
      expirationDate:
        type: string
      id:
        type: string
      land:
        allOf:
        - $ref: '#/definitions/main.Land'
        description: Relationships
      landID:
        type: string
      listings:
        items:
          $ref: '#/definitions/main.CreditListing'
        type: array
      totalCredits:
        type: number
      verificationID:
        type: string
      verificationStandard:
        type: string
      vintageYear:
        type: integer
    type: object
  main.CreateListingRequest:
    properties:
      carbonCreditsId:
        type: string
      maximumPurchase:
        type: number
      minimumPurchase:
        type: number
      pricePerCredit:
        type: number
      status:
        type: string
    type: object
  main.CreditListing:
    properties:
      carbonCredit:
        allOf:
        - $ref: '#/definitions/main.CarbonCredit'
        description: Relationships
      carbonCreditsID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      maximumPurchase:
        type: number
      minimumPurchase:
        type: number
      pricePerCredit:
        type: number
      status:
        type: string
      updatedAt:
        type: string
    type: object
  main.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  main.Land:
    properties:
      averageHumidity:
        type: number
      averageTemperature:
        type: number
      biomeType:
        type: string
      carbonCredits:
        items:
          $ref: '#/definitions/main.CarbonCredit'
        type: array
      certificationAuthority:
        type: string
      certificationDate:
        type: string
      createdAt:
        type: string
      description:
        type: string
      elevationMeters:
        type: number
      forestDensityPercentage:
        type: number
      id:
        type: string
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      ownerID:
        type: string
      seller:
        allOf:
        - $ref: '#/definitions/main.Seller'
        description: Relationships
      sizeSquareMeters:
        type: number
      soilType:
        type: string
      title:
        type: string
      treeSpecies:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      verificationStatus:
        type: string
    type: object
  main.Seller:
    properties:
      id:
        type: string
      lands:
        description: Relationships
        items:
          $ref: '#/definitions/main.Land'
        type: array
      userID:
        type: string
      verificationDate:
        type: string
      verificationStatus:
        type: string
    type: object
  main.UpdateListingRequest:
    properties:
      maximumPurchase:
        type: number
      minimumPurchase:
        type: number
      pricePerCredit:
        type: number
      status:
        type: string
    type: object
info:
  contact: {}
  title: Marketplace API
  version: "1.0"
paths:
  /api/market/active:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of all active listings with optional
        filters
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Minimum price filter
        in: query
        name: minPrice
        type: number
      - description: Maximum price filter
        in: query
        name: maxPrice
        type: number
      - description: Location filter
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Land'
            type: array
        "400":
          description: Invalid filters on request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
        "504":
          description: Request timed out
          schema:
            type: string
      summary: Get active listings
      tags:
      - listings
  /api/market/active/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific active listing by its ID
      parameters:
      - description: Listing ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Land'
        "400":
          description: Invalid listing ID
          schema:
            type: string
        "404":
          description: Listing not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
        "504":
          description: Request timed out
          schema:
            type: string
      summary: Get active listing by ID
      tags:
      - listings
  /api/market/health:
    get:
      description: Returns OK if the API is running
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Check API health
      tags:
      - Health
  /api/market/private:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of listings belonging to the authenticated
        seller
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Land'
            type: array
        "401":
          description: Unauthorized - Invalid or missing seller credentials
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get seller's private listings
      tags:
      - listings
    post:
      consumes:
      - application/json
      description: Creates a new credit listing for a seller's carbon credits
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: User Role (must be 'seller')
        in: header
        name: X-User-Role
        required: true
        type: string
      - description: Listing creation request
        in: body
        name: listing
        required: true
        schema:
          $ref: '#/definitions/main.CreateListingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created listing
          schema:
            $ref: '#/definitions/main.CreditListing'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create a new credit listing
      tags:
      - listings
  /api/market/private/{id}:
    delete:
      description: Deletes an existing credit listing owned by the seller
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: User Role (must be 'seller')
        in: header
        name: X-User-Role
        required: true
        type: string
      - description: Listing ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Listing not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Delete a credit listing
      tags:
      - listings
    get:
      consumes:
      - application/json
      description: Retrieves a specific listing belonging to the authenticated seller
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Listing ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Land'
        "400":
          description: Invalid listing ID
          schema:
            type: string
        "401":
          description: Unauthorized - Invalid or missing seller credentials
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Listing not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
        "504":
          description: Request timed out
          schema:
            type: string
      summary: Get seller's private listing by ID
      tags:
      - listings
    put:
      consumes:
      - application/json
      description: Updates an existing credit listing owned by the seller
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: User Role (must be 'seller')
        in: header
        name: X-User-Role
        required: true
        type: string
      - description: Listing ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Listing update request
        in: body
        name: listing
        required: true
        schema:
          $ref: '#/definitions/main.UpdateListingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated listing
          schema:
            $ref: '#/definitions/main.CreditListing'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Listing not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Update an existing credit listing
      tags:
      - listings
swagger: "2.0"